Add local mode and nonlocal field tag
Some of the fields may only be needed for over-the-network transfer,
but can be skipped when encoding the object locally. For instance, if
the object is stored in a blob field in the database, but some parts
of it also are stored in other columns in the same row, we can save
some space by omitting these fields in the blobs.

This is achieved by adding `nonlocal` field tag, `WithEncodeLocal()`
encoder option and `WithDecodeLocal()` decoder option. The encoder and
the decoder skip fields with `nonlocal` tag when they're in the local
mode.